made recipesmith code less monolithic
