+++ plugins/src/recipe_smith/structs/mod.rs
@@ -1,0 +1,17 @@
+mod crafter;
+mod ingredient;
+mod item;
+mod player_inventory;
+mod recipe;
+mod recipe_book;
+mod recipe_smith;
+mod storage_container;
+
+pub use crafter::Crafter;
+pub use ingredient::Ingredient;
+pub use item::Item;
+pub use player_inventory::PlayerInventory;
+pub use recipe::Recipe;
+pub use recipe_book::RecipeBook;
+pub use recipe_smith::RecipeSmith;
+pub use storage_container::StorageContainer;
